+++ src/solvers/dgmulti/types.jl
@@ -30,7 +30,9 @@ const DGMultiSBP{ApproxType, ElemType} =
 # these are necessary for pretty printing
 polydeg(dg::DGMulti) = dg.basis.N
 Base.summary(io::IO, dg::DG) where {DG <: DGMulti} = print(io, "DGMulti(polydeg=$(polydeg(dg)))")
-Base.real(rd::RefElemData{NDIMS, Elem, ApproxType, Nfaces, RealT}) where {NDIMS, Elem, ApproxType, Nfaces, RealT} = RealT
+
+# real(rd) is the eltype of the nodes `rd.r`.
+Base.real(rd::RefElemData) = eltype(rd.r)
 
 """
     DGMulti(; polydeg::Integer,
@@ -224,9 +226,10 @@ function DGMultiMesh(dg::DGMulti{NDIMS}; cells_per_dimension,
   end
 
   md = MeshData(vertex_coordinates, EToV, dg.basis)
-  if NDIMS == 1
-    md = StartUpDG.make_periodic(md, periodicity...)
-  else
+  if NDIMS == 1 && first(periodicity) == true
+    md = StartUpDG.make_periodic(md)
+  end
+  if NDIMS > 1
     md = StartUpDG.make_periodic(md, periodicity)
   end
   boundary_faces = StartUpDG.tag_boundary_faces(md, is_on_boundary)
@@ -318,13 +321,13 @@ function LinearAlgebra.mul!(b_in, A_kronecker::SimpleKronecker{2}, x_in)
 
   # copy `x_in` to `tmp_storage` to avoid mutating the input
   @assert length(tmp_storage) == length(x_in)
-  for i in eachindex(tmp_storage)
+  @turbo thread=true for i in eachindex(tmp_storage)
     tmp_storage[i] = x_in[i]
   end
   x = reshape(tmp_storage, n, n)
   b = reshape(b_in, n, n)
 
-  @turbo for j in 1:n, i in 1:n
+  @turbo thread=true for j in 1:n, i in 1:n
     tmp = zero(eltype(x))
     for ii in 1:n
       tmp = tmp + A[i, ii] * x[ii, j]
@@ -332,7 +335,7 @@ function LinearAlgebra.mul!(b_in, A_kronecker::SimpleKronecker{2}, x_in)
     b[i, j] = tmp
   end
 
-  @turbo for j in 1:n, i in 1:n
+  @turbo thread=true for j in 1:n, i in 1:n
     tmp = zero(eltype(x))
     for jj in 1:n
       tmp = tmp + A[j, jj] * b[i, jj]
@@ -340,7 +343,7 @@ function LinearAlgebra.mul!(b_in, A_kronecker::SimpleKronecker{2}, x_in)
     x[i, j] = tmp
   end
 
-  @turbo for i in eachindex(b_in)
+  @turbo thread=true for i in eachindex(b_in)
     b_in[i] = x[i]
   end
 
@@ -355,13 +358,13 @@ function LinearAlgebra.mul!(b_in, A_kronecker::SimpleKronecker{3}, x_in)
   n = size(A, 2)
 
   # copy `x_in` to `tmp_storage` to avoid mutating the input
-  for i in eachindex(tmp_storage)
+  @turbo thread=true for i in eachindex(tmp_storage)
     tmp_storage[i] = x_in[i]
   end
   x = reshape(tmp_storage, n, n, n)
   b = reshape(b_in, n, n, n)
 
-  @turbo for k in 1:n, j in 1:n, i in 1:n
+  @turbo thread=true for k in 1:n, j in 1:n, i in 1:n
     tmp = zero(eltype(x))
     for ii in 1:n
       tmp = tmp + A[i, ii] * x[ii, j, k]
@@ -369,7 +372,7 @@ function LinearAlgebra.mul!(b_in, A_kronecker::SimpleKronecker{3}, x_in)
     b[i, j, k] = tmp
   end
 
-  @turbo for k in 1:n, j in 1:n, i in 1:n
+  @turbo thread=true for k in 1:n, j in 1:n, i in 1:n
     tmp = zero(eltype(x))
     for jj in 1:n
       tmp = tmp + A[j, jj] * b[i, jj, k]
@@ -377,7 +380,7 @@ function LinearAlgebra.mul!(b_in, A_kronecker::SimpleKronecker{3}, x_in)
     x[i, j, k] = tmp
   end
 
-  @turbo for k in 1:n, j in 1:n, i in 1:n
+  @turbo thread=true for k in 1:n, j in 1:n, i in 1:n
     tmp = zero(eltype(x))
     for kk in 1:n
       tmp = tmp + A[k, kk] * x[i, j, kk]